+++ src/main/java/com/osroyale/Renderable.java
@@ -5,7 +5,7 @@ public class Renderable extends Cacheable {
 	public int modelHeight;
 	public VertexNormal aClass33Array1425[];
 
-	public void method443(int i, int j, int k, int l, int i1, int j1, int k1, int l1, int i2) {
+	public void method443(int i, int j, int k, int l, int i1, int j1, int k1, int l1, long i2) {
 		Model model = getRotatedModel();
 		if (model != null) {
 			modelHeight = model.modelHeight;
